+++ apig-20240327/ChangeLog.md
@@ -1,3 +1,9 @@
+2026-02-11 Version: 6.0.4
+- Update API CreateHttpApi: add request parameters body.belongGatewayId.
+- Update API CreateHttpApi: add request parameters body.dryRun.
+- Update API CreateHttpApi: add request parameters body.strategy.
+
+
 2026-02-10 Version: 6.0.3
 - Generated python 2024-03-27 for APIG.